fix(auth): break Principal/users/rbac import cycle (CodeQL #42-#46)
Phase 1 introduced ``app.auth.principal`` as a behaviour module —
it held the ``Principal`` dataclass *and* ``current_principal``
*and* the lazy import of ``get_user_permissions`` from rbac. That
created two cycles:

- principal -> users -> principal (Principal dataclass back-ref)
- principal -> rbac -> principal (current_principal <-> get_user_permissions)

The lazy imports kept the runtime working but CodeQL flagged the
graph as cyclic (alerts #42-#46 on PR #113).

Restructure: ``principal.py`` becomes a leaf with only the
dataclass + ``SCOPE_KEY`` (no app.auth.* imports). The
``current_principal`` resolver and the ``require_pat_scope`` helper
move into ``rbac.py`` alongside ``require_perm`` — they're tightly
coupled to it, and putting the trio in one module collapses the
graph cleanly. The private ``_sid_from_request`` helper moves with
them so rbac doesn't need to reach back into ``users`` twice.

``users.py`` now imports ``Principal`` and ``SCOPE_KEY`` at module
load, replacing the lazy import in ``current_user``'s body.

Public API unchanged: ``require_perm`` still returns ``User``,
``current_principal`` is still importable, ``Principal`` /
``SCOPE_KEY`` are still where ``pat_middleware.py`` expects them.

Tests: 252 passed (no behaviour change). Lint clean.
