+++ backend/app/auth/rbac.py
@@ -18,13 +18,19 @@ Usage:
 ``require_admin`` in ``app.auth.users`` is a convenience that checks
 for the ``admin`` RBAC role. Prefer ``require_perm`` for finer-grained
 gates so future roles compose cleanly.
+
+``current_principal`` and ``require_pat_scope`` live here too —
+they're tightly coupled to ``require_perm`` and keeping the trio
+in one module avoids the import cycle (``principal`` is a leaf with
+just the dataclass + ``SCOPE_KEY``).
 """
 from __future__ import annotations
 
-from fastapi import Depends, HTTPException, status
+from fastapi import Depends, HTTPException, Request, status
 from sqlalchemy import select
 from sqlalchemy.ext.asyncio import AsyncSession
 
+from app.auth.principal import SCOPE_KEY, Principal
 from app.auth.users import current_user
 from app.db import get_session
 from app.models.auth import User
@@ -69,10 +75,68 @@ async def current_user_permissions(
     return await get_user_permissions(session, user.id)
 
 
+async def current_principal(
+    request: Request,
+    user: User = Depends(current_user),
+    session: AsyncSession = Depends(get_session),
+) -> Principal:
+    """Resolve the principal for this request.
+
+    Two paths:
+
+    1. **PAT** — ``PATAuthMiddleware`` already authenticated the
+       bearer, intersected scopes with the user's current
+       permissions, and stashed a fully-formed ``Principal`` in
+       the scope slot. ``current_user`` short-circuits on the same
+       slot (returning the PAT user without the cookie / TOTP
+       gates), so by the time we run, ``user`` is the PAT user and
+       the scope slot still holds the pre-built principal — we
+       hand it back directly to preserve the scope-intersection
+       semantics.
+    2. **Cookie** — ``current_user`` enforced the totp + 2FA-
+       enrollment gates as usual. Resolve the user's full
+       permission set and wrap in a Principal for the caller.
+    """
+    pat_principal = request.scope.get(SCOPE_KEY)
+    if pat_principal is not None and isinstance(pat_principal, Principal):
+        # Re-use the middleware-built principal so PAT scope
+        # intersection wins over the user's full permission set.
+        return pat_principal
+
+    permissions = await get_user_permissions(session, user.id)
+    return Principal(
+        user=user,
+        permissions=frozenset(permissions),
+        auth_method="password",
+        token_id=None,
+        auth_session_id=_sid_from_request(request),
+    )
+
+
+def _sid_from_request(request: Request) -> str | None:
+    """Best-effort decode of the ``sid`` claim from the auth cookie.
+
+    Mirrors ``app.auth.users._sid_from_cookie`` but lives here so
+    ``current_principal`` doesn't need to reach into ``users``
+    twice (once via Depends, once for the helper).
+    """
+    import jwt as pyjwt
+
+    token = request.cookies.get("atrium_auth")
+    if not token:
+        return None
+    try:
+        data = pyjwt.decode(token, options={"verify_signature": False})
+        sid = data.get("sid")
+        return sid if isinstance(sid, str) else None
+    except Exception:
+        return None
+
+
 def require_perm(code: str):
     """Dependency factory: 403 unless the current principal has ``code``.
 
-    Now resolves through ``current_principal`` so PAT-authed and
+    Resolves through ``current_principal`` so PAT-authed and
     cookie-authed requests share one gate. ``permissions`` for a
     PAT request is the intersection of the token's scopes and the
     user's current permissions; for a cookie request it's the
@@ -83,8 +147,6 @@ def require_perm(code: str):
     Routes that need the auth method or token id can depend on
     ``current_principal`` directly.
     """
-    # Imported lazily to break the rbac → principal → rbac cycle.
-    from app.auth.principal import Principal, current_principal
 
     async def _dep(
         principal: Principal = Depends(current_principal),
@@ -95,6 +157,31 @@ def require_perm(code: str):
                 detail=f"permission '{code}' required",
             )
         return principal.user
+
+    return _dep
+
+
+def require_pat_scope(*scopes: str):
+    """Convenience factory for routes reachable *only* via a PAT
+    carrying one of the given scopes (PAT-management endpoints,
+    typically). Refuses cookie-authed requests with ``pat_required``.
+
+    Most routes should use ``require_perm(...)``. Reach for this
+    only when you need to refuse cookie auth explicitly.
+    """
+
+    async def _dep(principal: Principal = Depends(current_principal)) -> Principal:
+        if principal.auth_method == "password":
+            raise HTTPException(
+                status_code=status.HTTP_403_FORBIDDEN,
+                detail={"code": "pat_required"},
+            )
+        if not any(scope in principal.permissions for scope in scopes):
+            raise HTTPException(
+                status_code=status.HTTP_403_FORBIDDEN,
+                detail={"code": "insufficient_scope"},
+            )
+        return principal
 
     return _dep
